+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Version 66
 ----------
 *in development*
 
+* 🔧 When adding a show, do not fail if just Trakt info can not be loaded.
+
 #### 66.0.4 🧪
 *2022-12-16*